+++ source/tests/system/nidigital_driver_api_tests.cpp
@@ -1,13 +1,13 @@
 #include <gtest/gtest.h>
 
-#include "nidigitalpattern/nidigitalpattern_library.h"
-#include "nidigitalpattern/nidigitalpattern_service.h"
+#include "nidigital/nidigital_library.h"
+#include "nidigital/nidigital_service.h"
 
 namespace ni {
 namespace tests {
 namespace system {
 
-namespace digital = nidigitalpattern_grpc;
+namespace digital = nidigital_grpc;
 
 const int kDigitalDriverApiSuccess = 0;